Refactored the code in CIServer.java.
- separate the class WebhookJSONAnalyser from CIServer for maintainable structure.
